pyproject.toml
[tool.poetry]
name = "survmotion"
version = "0.1.0"
description = ""
authors = ["William Gomes F. Júnior <will.junior.go@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.8.1.78"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
